// src/users/dto/update-user.dto.ts
export class UpdateUserDto {
  name?: string;
  user_name?: string;
  college_number?: string;
  password?: string;
  telephone?: string;
  direction?: string;
  position?: string;
  email?: string;
  roles?: string;
  url_photo?: string;
  url_signature?: string;
  passwordSignature?: string;
  hide_user?: boolean;
}